34문제까지 완료

+++ src/com/javaex/practice/Ex33.java
@@ -11,8 +11,7 @@ public class Ex33 {
 		System.out.print("반지름: ");
 		double r = sc.nextDouble();
 		
-		double v = (4/3) * (3.14 * (r * r * r));
-		
+		double v = ((double)4/(double)3) * 3.14 * r * r * r;
 		System.out.println("구의 부피는: " + v + " 입니다.");
